hdmi_cec simplify atomic fallback #define

+++ esphome/components/hdmi_cec/hdmi_cec.h
@@ -159,7 +159,7 @@ class CECTransmit {
 #define ATOMIC_INCR(atom) ((atom)++)
 #else
 // This fallback is probably only for the older esp8266, which does not provide the atomic ++ operation
-#define ATOMIC_INCR(atom) ((atom) = (uint8_t) (atom) + 1)
+#define ATOMIC_INCR(atom) ((atom) = (atom) + 1)
 #endif
 
 class CECReceive {
